cms-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
bcrypt==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
Jinja2==3.1.2
reportlab==4.0.7
pandas==2.1.1
openpyxl==3.1.2